Finished Enemy's MovementModule
All 6 IQ levels are coded, added bonus "Chaoticness" parameter that describes how much the enemy is willing to undertake a process of changing platform or jumping onver a chasm (AI has a right to decide whether or not it "wants" to jump on other platform). 100 chaoticness = 100% willingness to change platforms 0 chaoticness = always stays on the same platform.